ws: Add msg queue to WS client using storage abstraction
JSONAsyncStorage is a simple abstraction on top of AsyncStorage to avoid
some boilerplate when handling JSON values.

If the WS connection is not open, messages will be added to the queue.
When the client re-connects, the messages are consumed via the send
method.

Note that `consume` empties the store before calling the consumer to
avoid a potential problem with side-effects from the consumer.

Github issue: https://github.com/coopcycle/coopcycle-app/issues/37