redux: Reorganise existing store and reducers
This re-organisation follows several principles that I tend to follow
when working on a redux app:

1. Normalising the shape of the state. This is based on one of the
recipes in the redux docs:
https://redux.js.org/recipes/structuring-reducers/normalizing-state-shape

The idea is to structure the store a little like a mini-database.

2. Using selectors. This is based on another recipe in the redux docs:
https://redux.js.org/recipes/computing-derived-data

The idea is (1) to decouple the shape of the state from your components
by using selector functions as a thin abstraction, and (2) to allow
those abstractions to take advantage of memoization to make computed
properties cheap.

3. Redux bundles. This is based on erikras' ducks proposal:
https://github.com/erikras/ducks-modular-redux

The idea here is to store all actions, reducers and selectors related to
a particular domain together

Github issue: https://github.com/coopcycle/coopcycle-app/issues/37