ws: Use redux actions instead of WS client on Tasks pages
The Registry no longer attaches listeners to the WS client because the
middleware takes care of that.

Many of the 'loading' states in the Tasks pages components are now
unnecessary.

In addition, fewer actions are required, as some were a result of WS
messages, which are now handled directly in the reducers.